% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/IlluminaMouse1Methylationanno.ilm10a1.mm10-package.R
\docType{package}
\name{IlluminaMouse1Methylationanno.ilm10a1.mm10-package}
\alias{IlluminaMouse1Methylationanno.ilm10a1.mm10}
\alias{IlluminaMouse1Methylationanno.ilm10a1.mm10-package}
\title{IlluminaMouse1Methylationanno.ilm10a1.mm10: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lihua Huang \email{atomnhuang@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
