% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IlluminaMouse1Methylationanno.R
\docType{data}
\name{IlluminaMouse1Methylationanno.ilm10a1.mm10}
\alias{IlluminaMouse1Methylationanno.ilm10a1.mm10}
\title{Data of annotation for illumina mouse methylation array on 02/11/2021}
\format{
A tibble with several sub DataFrames:
\describe{
\item{Locations}{probe locations}
\item{Month}{dbl Month price was recorded. Ranges from 1-12 for January - December}
\item{Manifest}{Manifest information about the probes}
\item{Islands.UCSC}{CpG island information}
\item{Other}{other annotation}
}
}
\source{
\url{https://github.com/atomnh/IlluminaMouseMethylationanno.ilm10a1.mm10}
}
\usage{
IlluminaMouse1Methylationanno.ilm10a1.mm10
}
\description{
Annotation object of IlluminaMethylationAnnotation.
}
\keyword{datasets}
